+++ src/user/user.controller.ts
@@ -6,70 +6,67 @@ import {
   Body,
   Delete,
   Put,
-  ParseIntPipe,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user-dto';
 import { UpdatePasswordDto } from './dto/update-password-dto';
+import {
+  ParseUUIDPipe,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
 
 @Controller('user')
 export class UserController {
   constructor(private userService: UserService) {}
 
   @Get()
-
-  // GET /user - get all users
-  // Server should answer with status code 200 and all users records
   getUsers() {
     return this.userService.getUsers();
   }
 
   @Get('/:userId')
-
-  //   GET /user/:id - get single user by id
-  //   Server should answer with status code 200 and and record with id === userId if it exists
-  //   Server should answer with status code 400 and corresponding message if userId is invalid (not uuid)
-  //   Server should answer with status code 404 and corresponding message if record with id === userId doesn't exist
-
-  // ParseIntPipe проверяет userId на Int при вводе в строке запроса
-  // getUser(@Param('userId', ParseIntPipe) userId: number) {
-  getUser(@Param('userId') userId: string) {
-    return this.userService.getUser(userId);
+  getUser(@Param('userId', ParseUUIDPipe) userId: string) {
+    const user = this.userService.getUser(userId);
+    if (!user) {
+      throw new NotFoundException(`User ID ${userId} not found`);
+    }
+    return user;
   }
-
-  //   POST /user - create user (following DTO should be used) CreateUserDto
-  //   interface CreateUserDto {
-  //     login: string;
-  //     password: string;
-  //   }
-  // Server should answer with status code 201 and newly created record if request is valid
-  // Server should answer with status code 400 and corresponding message if request body does not contain required fields
 
   @Post()
   createUser(@Body() createUserDto: CreateUserDto) {
     return this.userService.createUser(createUserDto);
   }
 
-  // PUT /user/:id - update user's password UpdatePasswordDto (with attributes):
-  // interface UpdatePasswordDto {
-  //   oldPassword: string; // previous password
-  //   newPassword: string; // new password
-  // }
-
   @Put('/:userId')
   updatePassword(
     @Body() updatePasswordDto: UpdatePasswordDto,
-    @Param('userId') userId: string,
+    @Param('userId', ParseUUIDPipe) userId: string,
   ) {
-    return this.userService.updatePassword(updatePasswordDto, userId);
+    const user = this.userService.getUser(userId);
+    if (!user) {
+      throw new NotFoundException(`User ID ${userId} not found`);
+    }
+
+    if (user.password !== updatePasswordDto.oldPassword) {
+      console.log(user.password);
+      console.log(updatePasswordDto.oldPassword);
+      throw new BadRequestException('Current password is incorrect');
+    }
+
+    this.userService.updatePassword(updatePasswordDto, userId);
+    return { message: 'Password updated successfully' };
   }
 
   @Delete('/:userId')
-  // DELETE /user/:id - delete user
-  // Server should answer with status code 204 if the record is found and deleted
-  // Server should answer with status code 400 and corresponding message if userId is invalid (not uuid)
-  // Server should answer with status code 404 and corresponding message if record with id === userId doesn't exist
-  deleteUser(@Param('userId') userId: string) {
-    return this.userService.deleteUser(userId);
+  deleteUser(@Param('userId', ParseUUIDPipe) userId: string) {
+    const user = this.userService.getUser(userId);
+    if (!user) {
+      throw new NotFoundException(`User ID ${userId} not found`);
+    }
+
+    this.userService.deleteUser(userId);
+    return { message: 'User deleted successfully' };
   }
 }
